Switched from self-written linear algebra to glam

+++ src/visual/rendering/polygon/polygon_filling_range.rs
@@ -5,12 +5,12 @@ use crate::math::interpolation::Interpolator;
 use super::polygon_interpolation_values::PolygonInterpolationValues;
 
 pub struct PolygonFillingRange<'a> {
-    pub range: Range<isize>,
-    pub interpolators: Vec<&'a (Interpolator<isize>, (PolygonInterpolationValues, PolygonInterpolationValues))>
+    pub range: Range<i32>,
+    pub interpolators: Vec<&'a (Interpolator<i32>, (PolygonInterpolationValues, PolygonInterpolationValues))>
 }
 
 impl<'a> Iterator for PolygonFillingRange<'a> {
-    type Item = isize;
+    type Item = i32;
     fn next(&mut self) -> Option<Self::Item> {
         self.range.next()
     }
